+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Switched to listr. Simplified code and more beautiful output.
 - Switched to execa. Should fix #30
 - Use ES2015. Dropped support for Node < 4.x
+- Support commands with arguments in the lint-staged config. Closes #47
+- Support binaries from $PATH. Closes #47
 
 # 2.0.2